requirements.txt
beautifulsoup4==4.11.2
fake_useragent==1.1.1
pandas==1.4.3
Pillow==9.4.0
requests==2.25.1
Unidecode==1.3.4
google_images_download
